Added support for CORS spring security to prevent data interference

+++ src/main/java/com/revature/web/ReviewController.java
@@ -2,6 +2,7 @@ package com.revature.web;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
 import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,10 +18,12 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
 import com.revature.model.Review;
+import com.revature.model.User;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.revature.model.Movies;
+import com.revature.model.Movies.Rating;
 import com.revature.service.MovieService;
 import com.revature.service.ReviewService;
 import com.revature.service.UserService;
@@ -78,6 +81,13 @@ public class ReviewController {
 				System.out.println("Request received for new movie: " + imdbID + ". Retrieving it from the OMDb API.");
 				String OMDbAPImovie = "http://www.omdbapi.com/?apikey=" + omdbApiKey + "&i=" + imdbID;
 				movie =  restTemplate.getForObject(OMDbAPImovie, Movies.class);
+				try {
+				List<Rating> RT = movie.getratings();
+				Rating score = RT.get(1);
+				movie.setRTscore(score.getValue());
+				} catch (IndexOutOfBoundsException e) {
+					movie.setRTscore("N/A");
+				}
 				this.movieService.insertMovie(movie);
 			}
 			return movie;
@@ -85,16 +95,24 @@ public class ReviewController {
 	
 	
 	@RequestMapping(value = "/savereviews", method = RequestMethod.POST)
-	public String saveReview(@RequestBody String r) {
+	public String saveReview(@RequestBody String r, HttpSession session) {
 		System.out.println("save reviews got called");
 		System.out.println(r);
 		String failure = "Failure";
-		//Object rparse = JSON.parse(r);
+		User currentlogin = new User();
+		
+        try {
+        	currentlogin = (User) session.getAttribute("USER");
+        	System.out.println("The current user is: " + currentlogin.getProfileid());
+        }
+        catch (NullPointerException e) {
+        	return "You are not currently logged in";
+        }
 		try {
 			String success = "Success";
-			
 			System.out.println("testing");
 			Review robj = new ObjectMapper().readValue(r, Review.class);
+			robj.setUserid(currentlogin.getProfileid());
 			System.out.println(robj.getRevdesc());
 			reviewService.insertReview(robj);
 			return success;
@@ -114,14 +132,30 @@ public class ReviewController {
     @ResponseBody
     @Transactional
     @PreAuthorize("#review.userId == principal.username")
-    public String deleteReview(@RequestBody String r) {
+    public String deleteReview(@RequestParam int revid, HttpSession session) {
+		User currentlogin = new User();
+	
         try {
-        	System.out.println("Now attempting to delete review number: " + r);
+        	currentlogin = (User) session.getAttribute("USER");
+        	System.out.println("The current user is: " + currentlogin.getProfileid());
+        }
+        catch (NullPointerException e) {
+        	return "This is not your review to delete";
+        }
+        try {
+        	System.out.println("Now attempting to delete review number: " + revid);
             String success = "Success";
-            Review robj = new ObjectMapper().readValue(r, Review.class);
-            reviewService.deleteReviewByRevid(robj.getRevid());
-            return success;
-        } catch (JsonProcessingException e) {
+            System.out.println("");
+            Review robj = reviewService.reviewsByOne(revid);
+            System.out.println("The id for the person that wrote this review is: " + robj.getUserid() + ". The id for the person sending this request is: "+ currentlogin.getProfileid());
+            if (robj.getUserid() == currentlogin.getProfileid()){
+            	reviewService.deleteReviewByRevid(robj.getRevid());
+                return success;
+            } else {
+            	return "This is not your review to delete";
+            }
+            
+        } catch (NullPointerException e) {
             e.printStackTrace();
             return "Failure";
         }
@@ -131,10 +165,16 @@ public class ReviewController {
 	
 	@RequestMapping(value = "/reviewsbyuser", method = RequestMethod.GET)
     public List<Review> reviewsByUser(@RequestParam String username){
+		try {
 		int userid = userService.findUseridByUsername(username);
 		System.out.println("User id is currently: " + userid);
         List<Review> reviewlist = reviewService.reviewsByUserid(userid);
         return reviewlist;
+		} catch (NullPointerException e) {
+			List<Review> emptylist = null;
+			return emptylist;
+		}
+		
 
     }
 	@RequestMapping(value = "/reviewsbymovie", method = RequestMethod.GET)
@@ -144,8 +184,8 @@ public class ReviewController {
 
     }
 	@RequestMapping(value = "/reviewsbyone", method = RequestMethod.GET)
-    public List<Review> reviewsByOne(@RequestParam int revid){
-        List<Review> reviewlist = reviewService.reviewsByOne(revid);
+    public Review reviewsByOne(@RequestParam int revid){
+        Review reviewlist = reviewService.reviewsByOne(revid);
         return reviewlist;
 
     }
